requirements.txt
boto3>=1.26.0
python-dotenv>=0.21.0
hdfs>=2.7.0
pyarrow>=10.0.0
cryptography>=38.0.0